Wikilog extension fixes: Bug 58154 - fix caching error Bug 56056 - do not handle post subarticles as posts themselves Bug 61175 - Default sort direction Bug 61149 - Show wikilog name in site feed Bug 51595 - Use DOM to extract summary from article content Bug 61632 - Make comment author signatures using generic mechanism Bug 61174 - Preview does not get parsed Bug 61174 - (prerequisite) $query->select(), $query->selectSQLText() Bug 62314 - Allow to reply from comment page Bug 62452 - Wikilog does not respect user timezones

+++ WikilogHooks.php
@@ -57,7 +57,7 @@ class WikilogHooks
 		if ( $title->isTalkPage() ) {
 			# ::WikilogCommentsPage::
 			# Invalidate cache of wikilog item page.
-			if ( $wi->getItemTitle()->exists() ) {
+			if ( $wi->getItemTitle() && $wi->getItemTitle()->exists() ) {
 				$wi->getItemTitle()->invalidateCache();
 				$wi->getItemTitle()->purgeSquid();
 			}
